Increase the size of ring buffer's byte counters to avoid overflow
Fix overflow of ring buffer counters which count the total bytes
over time i.e. total_bytes_written, total_bytes_read and
total_bytes_overwritten by increasing their size from u32 to u64.

CRs-Fixed: 2456561

+++ qcwcn/wifi_hal/ring_buffer.cpp
@@ -70,9 +70,9 @@ typedef struct ring_buf_cb {
     void (*threshold_cb)(void *);
     void *cb_ctx;
 
-    u32 total_bytes_written;
-    u32 total_bytes_read;
-    u32 total_bytes_overwritten;
+    u64 total_bytes_written;
+    u64 total_bytes_read;
+    u64 total_bytes_overwritten;
     u32 cur_valid_bytes;
     enum rb_bool threshold_reached;
 } rbc_t;
